# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentación de mi API con Echo.",
        "title": "API de Sliceflow",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1000",
    "basePath": "/",
    "paths": {
...
        "/hornero/authed/orders/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Muestra métricas y órdenes. Si es admin ve revenue, si no, ve $0.",
                "tags": [
                    "Production"
                ],
                "summary": "Dashboard Principal (Role-Based)",
                "responses": {}
            }
        },
...
        "/hornero/authed/orders/order": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Crear Orden de Trabajo",
                "parameters": [
                    {
                        "description": "Formulario Orden",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.CreateOrderDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/hornero/authed/orders/updord/{id}": {
            "put": {
                "description": "Permite editar detalles, asignar máquina o actualizar progreso (piezas hechas).",
                "tags": [
                    "Orders"
                ],
                "summary": "Actualizar Orden de Trabajo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la Orden",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos Nuevos",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.UpdateOrderDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "github_com_francotraversa_Sliceflow_internal_types.CreateOrderDTO": {
            "type": "object",
            "properties": {
                "client_name": {
                    "type": "string"
                },
                "deadline": {
                    "description": "String \"YYYY-MM-DD\"",
                    "type": "string"
                },
                "estimated_hours": {
                    "description": "input horas",
                    "type": "integer"
                },
                "estimated_minutes": {
                    "description": "input minutos",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "machine_id": {
                    "description": "Puede ser null (sin asignar)",
                    "type": "integer"
                },
                "material_id": {
                    "description": "\u003c--- El ID del combo box",
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "operator_id": {
                    "description": "El ID del usuario logueado o seleccionado",
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "priority": {
                    "description": "\"P1\", \"P2\", \"P3\"",
                    "type": "string"
                },
                "product_details": {
                    "type": "string"
                },
                "total_pieces": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_francotraversa_Sliceflow_internal_types.UpdateOrderDTO": {
            "type": "object",
            "properties": {
                "client_name": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "done_pieces": {
                    "type": "integer"
                },
                "estimated_hours": {
                    "type": "integer"
                },
                "estimated_minutes": {
                    "type": "integer"
                },
                "machine_id": {
                    "type": "integer"
                },
                "material_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "operator_id": {
                    "description": "Relaciones (Foreign Keys)",
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "priority": {
                    "description": "O int, segun como lo tengas",
                    "type": "string"
                },
                "product_details": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_pieces": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Escribí \"Bearer \" seguido de tu token JWT. Ejemplo: \"Bearer eyJhbG...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}